:construction: rearrange file lifecycle
Because of the way marked and hbs interact, compilation needs to happen in three stages -

1. Compile the section (file) hbs
2. Run the compiled hbs through the markdown renderer
3. Add the layout definition to the compiled markdown and compile it through the hbs engine again

The reason we need to go through this 3-step process is because marked escapes handlebars quotes, and templates have code indenting which causes marked to escape core code (i.e. nodes and text). So the easiest way to circumvent these is first focusing on the file, and then focusing on the layout