:bug: always expose instance to helpers
Nested helpers don't get access to the same context as root helpers. This brings up an issue for helpers which depend on instance-specific data (for example, the URL helper depends on the instance URL) - what if you want to call the helper in an #each loop? The fix is to create a proxy function which adds the instance object to the provided context, and then execute the function with `this` context